Original Command class is now SlashCommands. Added Command super class as well as User/Member commands
